feat(disapprove/resolve request): setup disapprove and resolve request

+++ client/js/logout.js
@@ -1,5 +1,7 @@
 const logout = document.getElementById('logout');
-logout.onclick = () => {
-  window.localStorage.clear();
-  window.location = baseUrl;
+if (logout) {
+  logout.onclick = () => {
+    window.localStorage.clear();
+    window.location = baseUrl;
+  };
 }
